requirements.txt
asgiref==3.3.1
auditok==0.1.5
autosub @ git+https://github.com/BingLingGroup/autosub.git@6a14c7b60d308e9f4c42ba9611d119346f0a9722
cachetools==4.2.4
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==2.0.7
colorama==0.4.4
deepspeech==0.9.3
Django==3.1.7
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
ffmpeg-normalize==1.22.4
ffmpeg-progress-yield==0.1.2
fuzzywuzzy==0.18.0
google-api-core==1.31.4
google-auth==1.35.0
google-cloud-speech==1.3.2
googleapis-common-protos==1.53.0
googletrans==4.0.0rc1
grpcio==1.42.0
grpcio-status==1.42.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.0.1
langcodes==3.3.0
libcst==0.3.21
lxml==4.6.4
mypy-extensions==0.4.3
numpy==1.17.3
packaging==21.3
pandas==1.2.3
Pillow==8.1.0
progressbar2==3.55.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
pyparsing==3.0.6
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pysubs2==1.3.0
python-dateutil==2.8.1
python-docx==0.8.11
python-Levenshtein==0.12.2
python-utils==2.5.6
pytz==2021.1
PyYAML==6.0
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.8.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.62.3
typing-extensions==4.0.0
typing-inspect==0.7.1
urllib3==1.26.7
wcwidth==0.2.5
webrtcvad==2.0.10
websocket-client==1.2.1
wincertstore==0.2
